+++ packages/component/src/provider/assetsProvider.tsx
@@ -5,7 +5,6 @@ import {
   createContext,
   useCallback,
   useContext,
-  useEffect,
 } from "react";
 import {
   useAccountInstance,
@@ -16,16 +15,11 @@ import {
   useWooCrossSwapQuery,
   useWooSwapQuery,
   useEventEmitter,
-  useSymbolsInfo,
   useWS,
 } from "@orderly.network/hooks";
 import { toast } from "@/toast";
 import { DepositAndWithdrawWithSheet } from "@/block/depositAndwithdraw/depositAndwithdraw";
-import {
-  capitalizeString,
-  transSymbolformString,
-} from "@orderly.network/utils";
-import { getOrderExecutionReportMsg } from "../block/orders/getOrderExecutionReportMsg";
+import { capitalizeString } from "@orderly.network/utils";
 
 export interface AssetsContextState {
   onDeposit: () => Promise<any>;
@@ -47,8 +41,6 @@ export const AssetsContext = createContext<AssetsContextState>(
 export const AssetsProvider: FC<PropsWithChildren> = (props) => {
   const { configStore } = useContext<any>(OrderlyContext);
   const account = useAccountInstance();
-
-  const ws = useWS();
 
   const onDeposit = useCallback(async () => {
     const result = await modal.show(DepositAndWithdrawWithSheet, {
@@ -115,36 +107,6 @@ export const AssetsProvider: FC<PropsWithChildren> = (props) => {
       }
     },
   });
-
-  const symbolsInfo = useSymbolsInfo();
-
-  useEffect(() => {
-    const unsubscribe = ws.privateSubscribe(
-      {
-        id: "executionreport",
-        event: "subscribe",
-        topic: "executionreport",
-        ts: Date.now(),
-      },
-      {
-        onMessage: (data: any) => {
-          // console.log("executionreport", data);
-          const { title, msg } = getOrderExecutionReportMsg(data, symbolsInfo);
-          if (title && msg) {
-            toast.success(
-              <div>
-                {title}
-                <br />
-                <div className="text-white/[0.54]">{msg}</div>
-              </div>
-            );
-          }
-          ee.emit("orders:changed", data);
-        },
-      }
-    );
-    return () => unsubscribe();
-  }, [symbolsInfo.symbol]);
 
   const { query: wooCrossSwapQuery } = useWooCrossSwapQuery();
   const { query: wooSwapQuery } = useWooSwapQuery();