+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,45 @@
 # examples
 
+## 0.1.64
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.92
+  - @orderly.network/core@0.0.87
+  - @orderly.network/hooks@0.0.95
+
+## 0.1.63
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.91
+  - @orderly.network/core@0.0.86
+  - @orderly.network/hooks@0.0.94
+
+## 0.1.62
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.90
+  - @orderly.network/core@0.0.85
+  - @orderly.network/hooks@0.0.93
+
+## 0.1.61
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.89
+  - @orderly.network/core@0.0.84
+  - @orderly.network/hooks@0.0.92
+
 ## 0.1.60
 
 ### Patch Changes