+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,29 @@
 # storybook
 
+## 1.0.74
+
+### Patch Changes
+
+- update
+
+## 1.0.73
+
+### Patch Changes
+
+- update
+
+## 1.0.72
+
+### Patch Changes
+
+- update
+
+## 1.0.71
+
+### Patch Changes
+
+- update
+
 ## 1.0.70
 
 ### Patch Changes