+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @orderly.network/cli
 
+## 0.0.32
+
+### Patch Changes
+
+- update
+
+## 0.0.31
+
+### Patch Changes
+
+- update
+
+## 0.0.30
+
+### Patch Changes
+
+- update
+
+## 0.0.29
+
+### Patch Changes
+
+- update
+
 ## 0.0.28
 
 ### Patch Changes