+++ packages/component/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @orderly.network/components
 
+## 0.0.92
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.87
+  - @orderly.network/types@0.0.87
+  - @orderly.network/utils@0.0.75
+
+## 0.0.91
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.86
+  - @orderly.network/types@0.0.86
+  - @orderly.network/utils@0.0.74
+
+## 0.0.90
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.85
+  - @orderly.network/types@0.0.85
+  - @orderly.network/utils@0.0.73
+
+## 0.0.89
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.84
+  - @orderly.network/types@0.0.84
+  - @orderly.network/utils@0.0.72
+
 ## 0.0.88
 
 ### Patch Changes