+++ packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,29 @@
 # eslint-config-custom
 
+## 0.0.28
+
+### Patch Changes
+
+- update
+
+## 0.0.27
+
+### Patch Changes
+
+- update
+
+## 0.0.26
+
+### Patch Changes
+
+- update
+
+## 0.0.25
+
+### Patch Changes
+
+- update
+
 ## 0.0.24
 
 ### Patch Changes