+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @orderly.network/hooks
 
+## 0.0.95
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@0.0.87
+  - @orderly.network/futures@0.0.89
+  - @orderly.network/net@1.0.87
+  - @orderly.network/types@0.0.87
+  - @orderly.network/utils@0.0.75
+
+## 0.0.94
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@0.0.86
+  - @orderly.network/futures@0.0.88
+  - @orderly.network/net@1.0.86
+  - @orderly.network/types@0.0.86
+  - @orderly.network/utils@0.0.74
+
+## 0.0.93
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.85
+  - @orderly.network/core@0.0.85
+  - @orderly.network/futures@0.0.87
+  - @orderly.network/types@0.0.85
+  - @orderly.network/utils@0.0.73
+
+## 0.0.92
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/core@0.0.84
+  - @orderly.network/futures@0.0.86
+  - @orderly.network/net@1.0.84
+  - @orderly.network/types@0.0.84
+  - @orderly.network/utils@0.0.72
+
 ## 0.0.91
 
 ### Patch Changes