+++ packages/react/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @orderly.network/react
 
+## 0.0.90
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.92
+  - @orderly.network/hooks@0.0.95
+
+## 0.0.89
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.91
+  - @orderly.network/hooks@0.0.94
+
+## 0.0.88
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.90
+  - @orderly.network/hooks@0.0.93
+
+## 0.0.87
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.89
+  - @orderly.network/hooks@0.0.92
+
 ## 0.0.86
 
 ### Patch Changes