+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @orderly.network/utils
 
+## 0.0.75
+
+### Patch Changes
+
+- update
+
+## 0.0.74
+
+### Patch Changes
+
+- update
+
+## 0.0.73
+
+### Patch Changes
+
+- update
+
+## 0.0.72
+
+### Patch Changes
+
+- update
+
 ## 0.0.71
 
 ### Patch Changes